requirements/ai_models.txt
transformers[ja, sklearn, tiktoken, serving, testing, sentencepiece, num2words, video, quality]>=4.54.1
x_transformers>=2.5.6
sentence_transformers>=5.0.0
datasets[audio]>=3.6.0
safetensors
tokenizers>=0.21.2
diffusers[docs]>=0.34.0
peft>=0.15.2
trl>=0.19.0
huggingface_hub[cli, hf_transfer, hf_xet, mcp]>=0.34.2
k_diffusion==0.0.12
hf_transfer>=0.1.9
hf_xet<2.0.0
fastrtc[vad,tts,stt]>=0.0.28
aiortc>=1.13.0
modelscope>=1.27.1
civitai-model-downloader
evaluate>=0.4.4
gradio[mcp]==5.39.0
gradio_webrtc
twilio
mcp[cli]
fastmcp
accelerate>=1.8.1
basicsr>=1.4.2
groq
spaces
facexlib>=0.3.0
gfpgan>=1.3.8
ultralytics==8.3.168
realesrgan @ git+https://github.com/xinntao/Real-ESRGAN.git
openai>=1.93.0
anthropic>=0.55.0
google-genai>=1.23.0
# xai-sdk>=1.0.0
ollama==0.5.1
lmstudio==1.4.1
unstructured[pdf]==0.18.3
langchain[community, anthropic, openai, google-genai, ollama, huggingface, perplexity, xai, mistralai]>=0.3.26
# langchain-chroma>=0.1.2
# langchain-neo4j>=0.4.0
qwen-vl-utils
qwen-omni-utils
# gptqmodel
gptqmodel @ git+https://github.com/ModelCloud/GPTQModel.git@29402342228414b81020a28a4c5b5828c1e76d59
jiwer
gtts>=2.5.4
chromadb
# outetts==0.3.3
pydub
samplerate
pyworld
ffmpeg-python
funasr
pyopenjtalk-plus
openai-whisper
Faster_whisper
wavmark
whisper-timestamped
gguf
llama-cpp-python
stable-diffusion-cpp-python
tipo-kgen

torch>=2.7.0
torchvision
torchaudio
protobuf==5.29.5
numpy>=2.0.2,<2.2.0
pandas
matplotlib
scikit-learn
fsspec==2025.3.0
numba>=0.60.0

# vllm; platform_system == 'Linux'
# llmcompressor; platform_system == 'Linux'